Move api WaterButlerMixin from nodes/views to base/views
