+++ api/nodes/views.py
@@ -45,7 +45,8 @@ from api.base.views import (
     BaseNodeLinksDetail,
     BaseNodeLinksList,
     LinkedNodesRelationship,
-    LinkedRegistrationsRelationship
+    LinkedRegistrationsRelationship,
+    WaterButlerMixin
 )
 from api.caching.tasks import ban_url
 from api.citations.utils import render_citation
@@ -91,7 +92,6 @@ from api.nodes.serializers import (
     NodeCitationSerializer,
     NodeCitationStyleSerializer
 )
-from api.nodes.utils import get_file_object
 from api.preprints.serializers import PreprintSerializer
 from api.registrations.serializers import RegistrationSerializer
 from api.users.views import UserMixin
@@ -168,39 +168,6 @@ class DraftMixin(object):
 
         self.check_object_permissions(self.request, draft.branched_from)
         return draft
-
-
-class WaterButlerMixin(object):
-
-    path_lookup_url_kwarg = 'path'
-    provider_lookup_url_kwarg = 'provider'
-
-    def get_file_item(self, item):
-        attrs = item['attributes']
-        file_node = BaseFileNode.resolve_class(
-            attrs['provider'],
-            BaseFileNode.FOLDER if attrs['kind'] == 'folder'
-            else BaseFileNode.FILE
-        ).get_or_create(self.get_node(check_object_permissions=False), attrs['path'])
-
-        file_node.update(None, attrs, user=self.request.user)
-
-        self.check_object_permissions(self.request, file_node)
-
-        return file_node
-
-    def fetch_from_waterbutler(self):
-        node = self.get_node(check_object_permissions=False)
-        path = self.kwargs[self.path_lookup_url_kwarg]
-        provider = self.kwargs[self.provider_lookup_url_kwarg]
-        return self.get_file_object(node, path, provider)
-
-    def get_file_object(self, node, path, provider, check_object_permissions=True):
-        obj = get_file_object(node=node, path=path, provider=provider, request=self.request)
-        if provider == 'osfstorage':
-            if check_object_permissions:
-                self.check_object_permissions(self.request, obj)
-        return obj
 
 
 class NodeList(JSONAPIBaseView, bulk_views.BulkUpdateJSONAPIView, bulk_views.BulkDestroyJSONAPIView, bulk_views.ListBulkCreateJSONAPIView, NodesFilterMixin, WaterButlerMixin):
